Add SettingsQR update info
* Move SettingsQR confirmation screen out of scan_screens.py
