+++ src/seedsigner/gui/screens/scan_screens.py
@@ -174,36 +174,3 @@ class ScanScreen(BaseScreen):
                     self.camera.stop_video_stream_mode()
                     break
 
-
-
-@dataclass
-class SettingsUpdatedScreen(ButtonListScreen):
-    config_name: str = None
-    title: str = "Settings QR"
-    is_bottom_list: bool = True
-
-    def __post_init__(self):
-        # Customize defaults
-        self.button_data = ["Home"]
-        self.show_back_button = False
-
-        super().__post_init__()
-
-        start_y = self.top_nav.height + 20
-        if self.config_name:
-            self.config_name_textarea = TextArea(
-                text=f'"{self.config_name}"',
-                is_text_centered=True,
-                auto_line_break=True,
-                screen_y=start_y
-            )
-            self.components.append(self.config_name_textarea)
-            start_y = self.config_name_textarea.screen_y + 50
-        
-        self.components.append(TextArea(
-            text="Settings imported successfully!",
-            is_text_centered=True,
-            auto_line_break=True,
-            screen_y=start_y
-        ))
-
